More Python3 fixes and epoloader fixes
Both epoloader and gpsstatus should now work in both
python2 and python3.  The other utilities still have
python3 issues but should still work with python2.

NOTE:  The EPO files currently available from MediaTek
contain 120 type II sets and are too large to be loaded
into the MT3339 in their entirety.  The maximum nunber
of sets seems to be 24 which you can set with the '-m'
option to epoloader.   This is just shy of one weeks
worth of data so you should schedule updates twice per
week.